refactor(voice): extract Windows SAPI backend to platform_speech/windows_speech.rs
Split the 971-line platform_speech.rs into a directory module. The
460-line inline `mod windows_speech` (SAPI 5.4 COM backend) moves
verbatim to platform_speech/windows_speech.rs; the module path is
unchanged (`platform_speech::windows_speech`), so every call site and
`use super::` resolves identically — a provably behavior-neutral move
(the cfg(windows) code can't be compiled on macOS, so this preserves it
exactly). mod.rs drops to 514 lines: cross-platform contract, macOS
backend, stub impl, and tests.

check: green (1480 tests, clippy/tsc/eslint clean).
